osc/rdma: Refactor btl wrapper code
Refactor wrapper calls around the btl atomic code into their own
header, removing them from the lock interface header.  Clean up
some resulting header dependency changes.

(cherry picked from commit 0394543786a55e90354829d80a5ba6eafa2c61b7)
